// client/src/registerPage/types/PostRegisterProps.ts
export type PostRegisterProps = {
  last_name: FormDataEntryValue | null;
  first_name: FormDataEntryValue | null;
  email: FormDataEntryValue | null;
  password1: FormDataEntryValue | null;
  password2: FormDataEntryValue | null;
  house_number: FormDataEntryValue | null;
  street_name: FormDataEntryValue | null;
  city: FormDataEntryValue | null;
  zip: FormDataEntryValue | null;
  state: FormDataEntryValue | null;
  country: FormDataEntryValue | null;
  birth_date: FormDataEntryValue | null;
  terms_and_conditions: FormDataEntryValue | null;
};